Beginning of script to generate obfuscated data
The intention here (not sure if it will work) will be to read in an ASCII art
image, and generate an obfuscated text body of that image by performing certain
numerical transformations on the character set representing the image. Should
also print the algorithm used to generate the image, or something like that.

Signed-off-by: Moses Mendoza <moses@puppet.com>