Add a character_metrics_to_location_array method
Given the data produced by the get_character_metrics method and an array
representing the set of unique characters in the content set we're dealing
with, generate a new set of data, an array of hashes containing :index and
:count keys where :index points to the location in the character set of the
given character, and :count points to the same count.

To make this a bit more readable, we update the original structure returned by
get_character_metrics to include descriptive key symbol values we can reference
later. This double the number of elements in each hash but its worth it.

Signed-off-by: Moses Mendoza <moses@puppet.com>