Add horrific multithreading
In a last-ditch effort to improve performance, split the computation of all
divmods into four sections and parallelize them with threads. This has
increased complexity and made it harder to understand, but our performance was
dying because we had to bump the ceiling under real world use given the total
indexes/counts.

Signed-off-by: Moses Mendoza <moses@puppet.com>