Add method to obtain the source integers
These are the integers that when, called divmod on with the second operand of
the solution, will return the result. These are the integers that we will need
to be the result of a preceding divmod operation.

ie, if one of these integers is x, then

i, x = i.divmod(n).

We will need to determine i, n such that x is the result. More complicated,
this is a psuedo-recursive operation. So the result of calling divmod on i will
be fed back into a new divmod.

Signed-off-by: Moses Mendoza <moses@puppet.com>